@@ -14,18 +14,18 @@
     "email": "coendunnink@gmail.com"
   },
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.0",
     "monolog/monolog": "^2.0",
-    "doctrine/orm": "^2.7"
+    "doctrine/orm": "^2.8"
   },
   "require-dev": {
     "phpunit/phpunit": "^9.5",
     "phpstan/phpstan": "^0.12",
     "phpstan/phpstan-doctrine": "^0.12",
     "phpstan/extension-installer": "^1.0",
-    "phpstan/phpstan-phpunit": "^0.12.8",
-    "friendsofphp/php-cs-fixer": "^2.16",
-    "phpstan/phpstan-strict-rules": "^0.12.2"
+    "phpstan/phpstan-phpunit": "^0.12",
+    "friendsofphp/php-cs-fixer": "^2.0",
+    "phpstan/phpstan-strict-rules": "^0.12"
   },
   "autoload": {
     "psr-4": {
@@ -34,8 +34,7 @@
   },
   "autoload-dev": {
     "psr-4": {
-      "SportsHelpers\\Tests\\": "tests/cases",
-      "SportsHelpers\\TestHelper\\": "tests/helpers"
+      "SportsHelpers\\Tests\\": "tests/cases"
     },
     "psr-0": {
       "": "tests/"
