@@ -4,10 +4,11 @@ declare(strict_types=1);
 
 namespace SportsHelpers\PouleStructure;
 
+use JetBrains\PhpStorm\Pure;
 use SportsHelpers\PouleStructure;
 
-class Balanced extends PouleStructure {
-
+class Balanced extends PouleStructure
+{
     public function __construct(int $nrOfPlaces, int $nrOfPoules)
     {
         $poules = [];
@@ -20,7 +21,8 @@ class Balanced extends PouleStructure {
         parent::__construct(array_values($poules));
     }
 
-    public function getRoundedNrOfPlacesPerPoule(bool $floor): int {
+    public function getRoundedNrOfPlacesPerPoule(bool $floor): int
+    {
         return $this->getNrOfPlacesPerPouleHelper($this->getNrOfPlaces(), $this->getNrOfPoules(), $floor);
     }
 
