@@ -3,6 +3,7 @@
 namespace SportsHelpers\Tests;
 
 use SportsHelpers\GameCalculatorDep;
+use SportsHelpers\GameMode;
 use SportsHelpers\SportBase;
 use SportsHelpers\SportConfig;
 
@@ -11,50 +12,48 @@ class SportConfigTest extends \PHPUnit\Framework\TestCase
     public function testCreation()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(2, $sportConfig->getNrOfFields() );
-        self::assertSame(2, $sportConfig->getNrOfGamePlaces() );
-        self::assertSame(1, $sportConfig->getGameAmount() );
+        self::assertSame(2, $sportConfig->getNrOfFields());
+        self::assertSame(2, $sportConfig->getNrOfGamePlaces());
+        self::assertSame(1, $sportConfig->getGameAmount());
     }
 
     public function testSport()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame($sport, $sportConfig->getSport() );
+        self::assertSame($sport, $sportConfig->getSport());
     }
 
     public function testToArray()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(["nrOfFields" => 2, "nrOfGamePlaces" => 2, "gameAmount" => 1 ], $sportConfig->toArray() );
+        self::assertSame(["gameMode" => GameMode::AGAINST, "nrOfFields" => 2, "nrOfGamePlaces" => 2, "gameAmount" => 1 ], $sportConfig->toArray());
     }
 
     public function testGetNrOfAgainstGames3Teams()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(3, $sportConfig->getNrOfGames($gameMode, 3) );
+        self::assertSame(3, $sportConfig->getNrOfGames(3));
     }
 
     public function testGetNrOfAgainstGames3TeamsGameAmount2()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 2;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(6, $sportConfig->getNrOfGames($gameMode, 3) );
+        self::assertSame(6, $sportConfig->getNrOfGames(3));
     }
 
     /**
@@ -63,61 +62,55 @@ class SportConfigTest extends \PHPUnit\Framework\TestCase
     public function testGetNrOfNotAgainstGames()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_TOGETHER;
+        $sport = new SportBase(GameMode::TOGETHER, 2);
         $gameAmount = 5;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(13, $sportConfig->getNrOfGames($gameMode, 5) );
+        self::assertSame(13, $sportConfig->getNrOfGames(5));
     }
 
     public function testGetNrOfAgainstGames5TeamsGameAmount2()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 2;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(20, $sportConfig->getNrOfGames($gameMode, 5) );
+        self::assertSame(20, $sportConfig->getNrOfGames(5));
     }
 
     public function testGetNrOfAgainstGames5TeamsGamePlaces4()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 4 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 4);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(15, $sportConfig->getNrOfGames($gameMode, 5) );
+        self::assertSame(15, $sportConfig->getNrOfGames(5));
     }
 
     public function testGetNrOfAgainstGames6TeamsGamePlaces4()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 4 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 4);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(45, $sportConfig->getNrOfGames($gameMode, 6) );
+        self::assertSame(45, $sportConfig->getNrOfGames(6));
     }
 
     public function testGetNrOfAgainstGames7TeamsGamePlaces4()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 4 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 4);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(105, $sportConfig->getNrOfGames($gameMode, 7) );
+        self::assertSame(105, $sportConfig->getNrOfGames(7));
     }
 
     public function testGetNrOfAgainstGames8TeamsGamePlaces4()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 4 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 4);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(210, $sportConfig->getNrOfGames($gameMode, 8) );
+        self::assertSame(210, $sportConfig->getNrOfGames(8));
     }
 
     /**
@@ -126,70 +119,63 @@ class SportConfigTest extends \PHPUnit\Framework\TestCase
     public function testGetNrOfAgainstGames8TeamsGamePlaces6()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 6 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 6);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(280, $sportConfig->getNrOfGames($gameMode, 8) );
+        self::assertSame(280, $sportConfig->getNrOfGames(8));
     }
 
     public function testGetNrOfNotAgainstGameRounds()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_TOGETHER;
+        $sport = new SportBase(GameMode::TOGETHER, 2);
         $gameAmount = 5;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(5, $sportConfig->getNrOfGameRounds($gameMode, 3) );
+        self::assertSame(5, $sportConfig->getNrOfGameRounds(3));
     }
 
     public function testGetNrOfAgainstGameRoundsEvenPlaces()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 5;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(15, $sportConfig->getNrOfGameRounds($gameMode, 4) );
+        self::assertSame(15, $sportConfig->getNrOfGameRounds(4));
     }
 
     public function testGetNrOfAgainstGameRoundsOddPlaces()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 5;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(25, $sportConfig->getNrOfGameRounds($gameMode, 5) );
+        self::assertSame(25, $sportConfig->getNrOfGameRounds(5));
     }
 
     public function testGetNrOfAgainstGameRoundsOddPlaces4GamePlaces()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 4 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 4);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(15, $sportConfig->getNrOfGameRounds($gameMode, 5) );
+        self::assertSame(15, $sportConfig->getNrOfGameRounds(5));
     }
 
     public function testNrOfGamesPerPlace()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_AGAINST;
+        $sport = new SportBase(GameMode::AGAINST, 2);
         $gameAmount = 1;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(4, $sportConfig->getNrOfGamesPerPlace($gameMode, 5) );
+        self::assertSame(4, $sportConfig->getNrOfGamesPerPlace(5));
     }
 
     public function testNrOfGamesPerPlaceTogether()
     {
         $nrOfFields = 2;
-        $sport = new SportBase( 2 );
-        $gameMode = SportConfig::GAMEMODE_TOGETHER;
+        $sport = new SportBase(GameMode::TOGETHER, 2);
         $gameAmount = 3;
         $sportConfig = new SportConfig($sport, $nrOfFields, $gameAmount);
-        self::assertSame(3, $sportConfig->getNrOfGamesPerPlace($gameMode, 5 ) );
+        self::assertSame(3, $sportConfig->getNrOfGamesPerPlace(5));
     }
 }
